Add `_is_abstract` method
Adds `_is_abstract` method to check if a node corresponds to an
abstract class. We can't instantiate objects from an abstract class,
so we shouldn't make them available in the `classes` dictionary.
Note: With static analysis via `AST` we can only infer whether a class
is abstract by heuristics, such as checking if the class inherits from
`ABC` or if it contains methods decorated with `@abstractmethod`.
This will not always work if the abstract classes are defined in a
more complex manner, such as assigning a different name to things like
`ABC` or `@abstractmethod` and using that name instead. A more
reliable way to determine if a class is abstract would be to do
runtime inspection with `inspect`, but it would be slower.